simpqs: store abs(X) in relations
We don't care about the sign, and this ensures we also remove duplicates
that differ only by sign. This is probably a bugfix, since enough of such
duplicates can cause the algorithm to fail.

This is enough to fix the example failure mentioned in 3c6cc9311c; it is
not yet known if it fixes all such failures.